+++ CHANGELOG.md
@@ -1,13 +1,21 @@
+# 0.2.0 (2019-04-xx)
+* [WIP] Added mongodb like state update support
+* Changed library format to UMD
+* Update dependencies
+
+# 0.1.3 (2019-03-19)
+* Update dependencies
+
 # 0.1.2 (2017-12-27)
 
 * Added hook methods: onDispatch() and onSetState()
 * Added method: log()
-* update dependencies
+* Update dependencies
 
 # 0.1.1 (2017-11-25)
 
 * Fixed compile error when strictNullChecks is false
-* update dependencies
+* Update dependencies
 
 # 0.1.0 (2017-09-01)